+++ component/meta/MetaTransit.js
@@ -29,4 +29,4 @@ module.exports = class MetaTransit extends Base {
 
 const ArrayHelper = require('areto/helper/ArrayHelper');
 const Forbidden = require('areto/error/ForbiddenHttpException');
-const Transition = require('evado-meta-document/workflow/Transition');
+const Transition = require('evado-meta-base/workflow/Transition');